Get rid of context passing (#29)
"Done" = compiles / runs existing tests, but have not even reviewed
myself.

This is the first big increment, wherein we get rid of context passing.
Start with the template app, it's better.
I think it leaves room for doing a lot of good things later.

Do bear in mind that the user will never be doing the root... some event
handler will. So no `dbos` instance in serverless, but something else.
And no need to ever make a root context, that'd always be done by HTTP
or the event receiver or whatever. In TS it was the testing runtime,
here we've really made it look like a bunch of functions which is not
realistic for an app. We'll keep evolving that... just think that if the
syntax got worse in some spot is that realistic once the rest of it is
done?