Move workflow stuff to core (#63)
Moves a few hundred lines out of DBOS facade into core.py

This is a checkpoint, we'd do the same to comm / tx / send / recv / etc.

Design note:
At the moment, a "DBOS" instance called 'dbos' is getting passed into
these functions. That object is really remembering the system db,
logger, registry of functions, etc... in the future I think we would
quickly change that to an instance of something other than `DBOS`... but
that'd be for later, probably part of the use of DBOS for decorators
etc.