Remove pesky stuff (#75)
The printouts on exit were covering cases that didn't warrant any
attention (such as importing dbos module without ever making a DBOS),
and generated confusion. Removed those as they weren't the interesting
ones.

Also removed an assert about how long something takes to run, as it is
flaky in the test runner, and was cut+paste from another test, so no
reason to have a redundant copy of that assert.