Detect functions passed instead of serializable arguments (#109)
For #111 

Detect functions passed as arguments / return values to DBOS functions,
and stop them.

Put a little more type info on WorkflowInput, and split the
serialization cases a little.

Generally, we may not want "unpicklable" items in the serializer, but
Kafka currently uses it.

Editorial comment, if this code is catching something, I wonder if the
end developer is using mypy or similar. Hate to preach, but they're
making a big mistake if they don't... are we adequately stressing this?