Prevent different workflows starting with the same ID (#185)
This fix is similar to the one in TypeScript:
https://github.com/dbos-inc/dbos-transact-ts/pull/726

- This PR mainly addresses the issue where different workflows started
with the same workflow ID breaks the OAOO guarantees. The solution is to
check the recorded function name, class name, and config name (if any)
before proceeding to start the workflow. If there's any difference,
throw a `DBOSConflictingWorkflowError` with a clear error message.
- Plus, if multiple workflow started within the SetWorkflowID block,
only the first workflow will be assigned with the specified workflow ID;
subsequent workflows will use a generated workflow ID. Make sure we have
a clear warning message for users.
- This PR also checks the recorded input versus the provided input, and
the provided queue name versus the original queue name, and print a
warning message if they don't match.
- It also fixes the Kafka consumer workflow IDs and adds a way to
generate a group ID based on the function name and topics. This should
match the TS implementation.