remove explicit notion of recovery (#198)
This PR simplifies the logic for putting a workflow in the dead letter
queue. Specifically, it changes the meaning of the `recovery_attempt`
column in the `dbos.workflow_status` table to `attempts`*.

The default of the column is now 1.

`update_workflow_status` just verifies that (number of attempts) <=
max_retries before placing a workflow in the DLQ.

[*] This PR does not change the column name, to facilitate backward
compatibility.