Fix race conditions with clear_queue_assignment (#232)
It is possible for a task to be sent back to the `QUEUED` state if
Transact crashes between removing the task from the queue, after
completion, and flushing the workflow output buffer.

At restart, the queue will not pickup on the workflow (because it has
been dequeued), and we skip the workflow during recovery -- so the
status is never set to `SUCCESS`.

This PR:
- Has `clear_queue_assignment` check whether the workflow is still in
the queue (checking the affected `rowcount` from the workflow queue
clear)
- Has `clear_queue_assignment` return whether it did find a workflow to
clear in the queue. If not, the recovery logic considers the workflow is
not a queue task anymore and needs to be re-executed immediately