Background Event Loop (#292)
Create a single event loop to which DBOS submits any coroutines that are
not started from within an event loop. In particular, coroutines
submitted to queues (such as from scheduled workflows) run on this event
loop. This fixes an issue where coroutines are submitted to event loops
on threads which are later killed when the function that started the
thread completes, preventing the coroutines from finishing.