Trim Dependencies (#464)
This PR eliminates unnecessary dependencies of DBOS. It includes
breaking changes, such as the removal of the deprecated Alembic-based
migrations and making OpenTelemetry optional and disabled by default. It
also removes support for Python 3.9, which is reaching end of life. It
will be released as a new major version of DBOS (2.0).


Eliminated dependencies:

- jsonschema
- alembic
- typing-extensions
- jsonpickle
- fastapi[standard]
- tomlkit
- docker
- cryptography
- rich
- pyjwt

Additionally the three OTLP dependencies (opentelemetry-api,
opentelemetry-sdk, opentelemetry-exporter-otlp-proto-http) are now
optional and typer is replaced with typer-slim.